requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-extensions==3.2.1
django-localflavor==4.0
django-phonenumber-field==7.1.0
idna==3.4
phonenumbers==8.13.11
phonenumberslite==8.13.10
Pillow==9.5.0
python-decouple==3.8
python-dotenv==1.0.0
python-stdnum==1.18
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15